// CPL/admin-backend/libs/import-file/src/import-file.enum.ts
export enum BALANCE_IMPORT_FILE_STATUS {
  PROCESSING = 1,
  SUCCESS = 2,
  CANCELLED = 3,
}

export enum BALANCE_IMPORT_ROW_STATUS {
  PROCESSING = 1,
  SUCCESS = 2,
  CANCELLED = 3,
  ERROR = 4,
}

export enum IMPORT_FILE_ROW_ERRORS {
  MAX_LINE_AMOUNT = 'MAX_LINE_AMOUNT',
  MAX_FILE_AMOUNT = 'MAX_FILE_AMOUNT',
  USER_NOT_FOUND = 'USER_NOT_FOUND',
  USER_IS_BANNED = 'USER_IS_BANNED',
  USER_NOT_KYC = 'USER_NOT_KYC',
  USER_IS_DELETED = 'USER_IS_DELETED',
  CURRENCY_NOT_SETTING = 'CURRENCY_NOT_SETTING',
  CURRENCY_NOT_AVAILABLE = 'CURRENCY_NOT_AVAILABLE',
  USER_IS_PENDING_DELETED = 'USER_IS_PENDING_DELETED',
}

export enum IMPORT_FILE_ERRORS {
  AMOUNT_ERROR = 'AMOUNT_ERROR',
  INVALID_FILE_FORMAT = 'INVALID_FILE_FORMAT',
}

export const ITEMS_PER_PAGE = 100
export const USER_ID_DEFAULT = '-1'